requirements/local.txt
-r ./base.txt
factory-boy==2.11.1
Faker==1.0.6
flake8==3.7.7
pytest==4.4.2
pytest-django==3.4.8
